RIA-7175 Refactor of package
